refactor(web): move file position
