+++ src/web/components/modals/ChangePassword.scss
@@ -1,5 +1,5 @@
-@import "../../shared/utils/color";
-@import "../../shared/utils/mixins";
+@import "../../../shared/utils/color";
+@import "../../../shared/utils/mixins";
 
 .change-password {
   display: flex;
